+++ daemon/src/ctrl_anime/zbus.rs
@@ -1,13 +1,11 @@
 use std::sync::{Arc, Mutex};
 
 use async_trait::async_trait;
-use log::warn;
 use rog_anime::{
     usb::{pkt_for_apply, pkt_for_set_boot, pkt_for_set_on},
     AnimeDataBuffer, AnimePowerStates,
 };
 use zbus::{dbus_interface, Connection, SignalContext};
-use zvariant::ObjectPath;
 
 use std::sync::atomic::Ordering;
 
@@ -19,18 +17,7 @@ pub struct CtrlAnimeZbus(pub Arc<Mutex<CtrlAnime>>);
 #[async_trait]
 impl crate::ZbusAdd for CtrlAnimeZbus {
     async fn add_to_server(self, server: &mut Connection) {
-        server
-            .object_server()
-            .at(
-                &ObjectPath::from_str_unchecked("/org/asuslinux/Anime"),
-                self,
-            )
-            .await
-            .map_err(|err| {
-                warn!("CtrlAnimeDisplay: add_to_server {}", err);
-                err
-            })
-            .ok();
+        Self::add_to_server_helper(self, "/org/asuslinux/Anime", server).await;
     }
 }