+++ daemon/src/lib.rs
@@ -34,8 +34,10 @@ use std::time::Duration;
 use crate::error::RogError;
 use async_trait::async_trait;
 use config::Config;
+use log::warn;
 use smol::{stream::StreamExt, Executor, Timer};
 use zbus::Connection;
+use zvariant::ObjectPath;
 
 pub static VERSION: &str = env!("CARGO_PKG_VERSION");
 
@@ -46,12 +48,30 @@ pub trait Reloadable {
 #[async_trait]
 pub trait ZbusAdd {
     async fn add_to_server(self, server: &mut Connection);
+
+    async fn add_to_server_helper(
+        iface: impl zbus::Interface,
+        path: &str,
+        server: &mut Connection,
+    ) {
+        server
+            .object_server()
+            .at(&ObjectPath::from_str_unchecked(path), iface)
+            .await
+            .map_err(|err| {
+                warn!("{}: add_to_server {}", path, err);
+                err
+            })
+            .ok();
+    }
 }
 
 /// Set up a task to run on the async executor
 #[async_trait]
 pub trait CtrlTask {
-    async fn create_task(&self, executor: &mut Executor) -> Result<(), RogError>;
+    /// Implement to set up various tasks that may be required, using the `Executor`.
+    /// No blocking loops are allowed, or they must be run on a separate thread.
+    async fn create_tasks(&self, executor: &mut Executor) -> Result<(), RogError>;
 
     /// Create a timed repeating task
     async fn repeating_task(